moved swagger from handlers to v1
