+++ app/services/kitchen/api/v1/swagger.go
